Change material card formatter to work without number.
Need to hash material cards when using activation control (ACT)
card with a full core, because the ACT card requires allocation of an
array that is (nCells x nIsotopeMats*3), where nIsotopeMats is the
the number of all isotope-material combinations occuring in the input.
(i.e., 100 materials with 20 isotopes each => nIsotopeMats = 2000). This
scales very quickly for a third-core or full-core problem, and the array
can easily exceed 100 GB in size. Even without heterogeneous pin
representation, it would likely be dozens of GB.

For a fresh core, we can significantly reduce the number of materials in
the problem (probably 20-30, depending on the number of assembly types),
but when depletion comes into play the material number will explode
unless some type of region-averaged depletion is implemented. But we'll
cross that bridge when (or if) we get there.