Set GeomType to HEX when creating Hex grid with fromPitch().
fromPitch() returns a HexGrid, but it does not set the GeomType or the
SymmetryType of the grid. The GeomType should certainly be HEX. Whether
a default symmetry type should be set is debatable, so it is left blank
for the moment. If the user requests the symmetry from this grid, there
will be an error.