Add nozzleType parameter for assembly.
Provide more flexibility to orificing by adding a nozzleType parameter.
The orificer already works with this in principle for the LTA
assemblies, but the nozzle types are hard-coded as 1: regular, 2: LTA.
With this commit, the user can deifne an arbitrary number of nozzle
types by setting it for each assembly in the blueprints.