Fix for crWorth error power mismatch.
Not clear whether this is the actual fix for the bug.
crStoreIntermediateDatabases: false also appears to be helpful.

The bug is that sometimes when running crWorth cases, Nala will crash on
the uniform mesh converter in _checkConversion() when the reactor power
does not equal the specified power. The hypothesis is that this happens
when a third-core model is expanded to full core for an individual rod
worth calculation or bank worth -1 rod calculation. This asymmetrical
core produces an asymmetrical power shape; when
geometryConverter.restoreGeometry() is then run to restore third-core
geometry, the resulting reactor does not have the correct power (1/3 of
specified power). This is interpreted by the code as an unsuccessful
uniform mesh conversion, but it is actually just the result of
incorrectly "restoring" third core symmetry to a model that does not
have symmetry. Instead of trying to map these full core parameters back
to the third core, we should just leave them alone and not map new
values.