Conserve mass at component level for a.setBlockMesh()
a.setBlockMesh() had an option, `conserveMassFlag="auto"`, which would
try to automatically conserve mass of the fuel while allowing other
components to change mass with changing block heights. This conservation
was based upon getting the nuclides associated with the fuel component,
and then scaling the density of those nuclides based on the change in
block height using b.setNumberDensities().

This method does not work when the fuel has nuclides that are also in
other components of the block, like cladding or structure. In this case,
the mass of those components will also be conserved, which is not the
desired behavior.

Instead, a.setBlockMesh() needs to conserve mass by setting number
densities directly on the components where mass needs to be conserved,
instead of setting the densities on the full block.